latest network code. Network is under development

+++ src/main/java/edu/chl/Game/view/screens/OptionView.java
@@ -134,7 +134,8 @@ public class OptionView extends AbstractMenuScreen {
 		
 		//Buttons and Labels for submenu Sound
 		Label soundLabel = new Label("Sound:", skin);
-		Label soundStatusLabel = new Label(Integer.toString(Sound.getCurrentVolume()), skin);
+		//int currentVol = (int)Sound.getCurrentVolume();
+		Label soundStatusLabel = new Label(Integer.toString((int)Sound.getCurrentVolume()), skin);
 		
 		TextButton soundIncrease = new TextButton("+", skin);
 		soundIncrease.addListener(mouseInput);
